+++ src/features/my-page/pages/shop-setting-page.tsx
@@ -1,4 +1,4 @@
-import { Aside } from '../../../../shared/components/layout';
+import { Aside } from '../../../shared/components/layout';
 
 import { ToggleSwitch } from '../components/ui/toggleswitch';
